Move log_training_data() to parent class
